pyproject.toml
[tool.poetry]
name = "glaucofilho-rinha-de-backend-2024-q1"
version = "0.1.0"
description = ""
authors = ["glaucofilho <glaucowow@hotmail.com.br>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
pydantic-settings = "^2.2.0"
uvicorn = "^0.27.1"
SQLAlchemy = "^2.0.27"
asyncpg = "^0.29.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
